repo-guardian/requirements.txt
openai
PyGithub
gitpython
pyyaml